refactor: re-use Button code
